chore(sdk-coin-sol): move account-lib code
- moves sol code from account-lib

Ticket: BG-48156
