+++ modules/sdk-coin-sol/package.json
@@ -41,7 +41,13 @@
     ]
   },
   "dependencies": {
-    "@bitgo/sdk-core": "^1.0.1"
+    "@bitgo/sdk-core": "^1.0.1",
+    "@bitgo/statics": "^6.17.0",
+    "@solana/spl-token": "0.1.8",
+    "@solana/web3.js": "1.31.0",
+    "bignumber.js": "^9.0.0",
+    "bs58": "^4.0.1",
+    "tweetnacl": "^1.0.3"
   },
   "devDependencies": {
     "@bitgo/sdk-api": "^1.0.1",
