+++ modules/sdk-coin-sol/src/index.ts
@@ -1,2 +1,1 @@
-export * from './sol';
-export * from './tsol';
+export * from './lib';
