+++ modules/sdk-coin-sol/src/lib/constants.ts
