+++ modules/sdk-coin-sol/src/lib/keyPair.ts
