+++ modules/sdk-coin-sol/tsconfig.json
@@ -18,6 +18,9 @@
     },
     {
       "path": "../sdk-test"
+    },
+    {
+      "path": "../statics"
     }
   ]
 }
